requirements.txt
apache-airflow
requests
psycopg2
psycopg2-binary
sqlalchemy
pandas
PyYAML
requests